Add new directive to allow using "formulas" for fields
The new directive adds the possibility describe the calculated value of
a field by using a "formula" (which is an angular expression) much like
in spreadsheet software.

It is possible to define fields to be watched and which expression
should be evaluated. This is more flexible than defining multiple
"onChange"-properties for different fields and is directly related to
the field on which it is defined.