f|Add attributes and typeahead builders
Using the attributes builder, it is possible to append any attribute to
the input-field of a template. This even enables the use of custom
directives, as the templates are later ().

The typeahead builder enables the use of the uib-typeahead directive
from angular-bootstrap. A typeahead function must be supplied in the
form options.